Add migration for the cache lookup table
